+++ source/neurons/decision_making.hpp
@@ -24,7 +24,7 @@ namespace hummus {
 	class Decision_Making : public Neuron {
 	public:
 		// ----- CONSTRUCTOR AND DESTRUCTOR -----
-        Decision_Making(int _neuronID, int _layerID, int _sublayerID, int _rf_id,  std::pair<int, int> _xyCoordinates, std::string _classLabel="", int _refractoryPeriod=0, double _traceTimeConstant=20, double _threshold=-50, double _restingPotential=-70) :
+        Decision_Making(int _neuronID, int _layerID, int _sublayerID, int _rf_id,  std::pair<int, int> _xyCoordinates, std::string _classLabel="", int _refractoryPeriod=0, float _traceTimeConstant=20, float _threshold=-50, float _restingPotential=-70) :
                 Neuron(_neuronID, _layerID, _sublayerID, _rf_id, _xyCoordinates, _refractoryPeriod, 200, 10, _traceTimeConstant, _threshold, _restingPotential, _classLabel) {
 
             // DecisionMaking neuron type = 2 for JSON save
@@ -47,7 +47,7 @@ namespace hummus {
             }
         }
 
-        virtual void update(double timestamp, Synapse* s, Network* network, double timestep, spike_type type) override {
+        virtual void update(double timestamp, Synapse* s, Network* network, float timestep, spike_type type) override {
 
             // checking if the neuron is in a refractory period
             if (timestamp - previous_spike_time >= refractory_period) {
@@ -103,8 +103,8 @@ namespace hummus {
             }
         }
 
-        virtual double share_information() override {
-            return static_cast<double>(intensity);
+        virtual float share_information() override {
+            return static_cast<float>(intensity);
         }
 
         // write neuron parameters in a JSON format
