+++ source/random_distributions/cauchy.hpp
@@ -20,16 +20,16 @@ namespace hummus {
         
 	public:
 		// ----- CONSTRUCTOR AND DESTRUCTOR -----
-        Cauchy(double weight_location=1, double weight_scale=0, double delay_location=0, double delay_scale=0) {
+        Cauchy(float weight_location=1, float weight_scale=0, float delay_location=0, float delay_scale=0) {
 
             // randomising weights and delays
             std::random_device device;
             random_engine = std::mt19937(device());
-            delay_random = std::cauchy_distribution<double>(delay_location, delay_scale);
-            weight_random = std::cauchy_distribution<double>(weight_location, weight_scale);
+            delay_random = std::cauchy_distribution<float>(delay_location, delay_scale);
+            weight_random = std::cauchy_distribution<float>(weight_location, weight_scale);
         }
 		
-        std::pair<double, double> operator()(int x, int y, int depth) {
+        std::pair<float, float> operator()(int x, int y, int depth) {
 			return std::make_pair(weight_random(random_engine), std::abs(delay_random(random_engine)));
         }
         
@@ -37,8 +37,8 @@ namespace hummus {
         
         // ----- IMPLEMENTATION VARIABLES -----
         std::mt19937                     random_engine;
-        std::cauchy_distribution<double> delay_random;
-        std::cauchy_distribution<double> weight_random;
+        std::cauchy_distribution<float> delay_random;
+        std::cauchy_distribution<float> weight_random;
 	};
 }
